bench: Wire up jtag interface and extend  image makefile
I am working on testing openocd with the de0 nano and confirming
the sim works as expected is a first step.

The jtag debug interface was not working as expected as it was
never wired up correctly, wire it up and it seems to work fine
after that.

Add a few comments to the Makefile for spi images, I tested with
other images that do not just exit right away.